Added hylide for shaders in haskell and clay for css in haskell.
The idea is to remove the custom node based client app in hylide and replace it with a reflex dom version.
This will become a haskell based framework for interactive shader work.

For now interactivity is done in vscode with ghcid, but it should hopefully be possible to directly edit haskell code in the browser in the future.